+++ spec/rubocop/options_spec.rb
@@ -156,14 +156,6 @@ Usage: rubocop [options] [file1, file2, ...]
   end
 
   unless RuboCop::Version::STRING.start_with?('0')
-    describe '-e/--emacs option' do
-      it 'is dropped in RuboCop 1.0.0' do
-        # This spec can be removed once the option is dropped.
-        expect { options.parse(['--emacs']) }
-          .to raise_error(OptionParser::InvalidOption)
-      end
-    end
-
     describe '-s/--silent option' do
       it 'raises error in RuboCop 1.0.0' do
         # This spec can be removed